+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # mailing
 
+## 0.6.2
+
+### Patch Changes
+
+- export render from mailing-core
+
+## 0.6.1
+
+### Patch Changes
+
+- bugfix js support
+
 ## 0.6.0
 
 ### Minor Changes