+++ packages/web/CHANGELOG.md
@@ -1,5 +1,17 @@
 # web
 
+## 0.2.2
+
+### Patch Changes
+
+- export render from mailing-core
+
+## 0.2.1
+
+### Patch Changes
+
+- bugfix js support
+
 ## 0.2.0
 
 ### Minor Changes